+++ plugin/providers/phpipam/data_source_phpipam_address_test.go
@@ -26,6 +26,25 @@ data "phpipam_address" "address_by_description" {
 }
 `
 
+const testAccDataSourcePHPIPAMAddressCustomFieldConfig = `
+resource "phpipam_address" "address" {
+  subnet_id   = 3
+  ip_address  = "10.10.1.10"
+  description = "Terraform test address (custom fields)"
+  hostname    = "tf-test.cust1.local"
+
+  custom_fields = {
+    CustomTestAddresses = "terraform-test"
+  }
+}
+
+data "phpipam_address" "custom_search" {
+  subnet_id                 = "${phpipam_address.address.subnet_id}"
+  custom_field_filter_key   = "CustomTestAddresses"
+  custom_field_filter_value = ".*terraform.*"
+}
+`
+
 func TestAccDataSourcePHPIPAMAddress(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
@@ -39,6 +58,25 @@ func TestAccDataSourcePHPIPAMAddress(t *testing.T) {
 					resource.TestCheckResourceAttr("data.phpipam_address.address_by_address", "description", "Gateway"),
 					resource.TestCheckResourceAttr("data.phpipam_address.address_by_hostname", "ip_address", "10.10.1.3"),
 					resource.TestCheckResourceAttr("data.phpipam_address.address_by_description", "ip_address", "10.10.1.4"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccDataSourcePHPIPAMAddress_CustomField(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDataSourcePHPIPAMAddressCustomFieldConfig,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.phpipam_address.custom_search", "subnet_id", "3"),
+					resource.TestCheckResourceAttr("data.phpipam_address.custom_search", "ip_address", "10.10.1.10"),
+					resource.TestCheckResourceAttr("data.phpipam_address.custom_search", "description", "Terraform test address (custom fields)"),
+					resource.TestCheckResourceAttr("data.phpipam_address.custom_search", "hostname", "tf-test.cust1.local"),
+					resource.TestCheckResourceAttr("data.phpipam_address.custom_search", "custom_fields.CustomTestAddresses", "terraform-test"),
 				),
 			},
 		},