+++ plugin/providers/phpipam/resource_phpipam_address.go
@@ -1,6 +1,9 @@
 package phpipam
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/paybyphone/phpipam-sdk-go/phpipam"
 )
@@ -31,6 +34,23 @@ func resourcePHPIPAMAddressCreate(d *schema.ResourceData, meta interface{}) erro
 		return err
 	}
 
+	// If we have custom fields, set them now. We need to get the IP address's ID
+	// beforehand.
+	if customFields, ok := d.GetOk("custom_fields"); ok {
+		addrs, err := c.GetAddressesByIP(in.IPAddress)
+		if err != nil {
+			return fmt.Errorf("Could not read IP address after creating: %s", err)
+		}
+
+		if len(addrs) != 1 {
+			return errors.New("IP address either missing or multiple results returned by reading IP after creation")
+		}
+
+		if _, err := c.UpdateAddressCustomFields(addrs[0].ID, customFields.(map[string]interface{})); err != nil {
+			return err
+		}
+	}
+
 	return dataSourcePHPIPAMAddressRead(d, meta)
 }
 
@@ -42,6 +62,10 @@ func resourcePHPIPAMAddressUpdate(d *schema.ResourceData, meta interface{}) erro
 	in.IPAddress = ""
 	in.SubnetID = 0
 	if _, err := c.UpdateAddress(in); err != nil {
+		return err
+	}
+
+	if err := updateCustomFields(d, c); err != nil {
 		return err
 	}